+++ src/systems/move_system.rs
@@ -43,40 +43,39 @@ impl<'a> System<'a> for MoveSystem {
             let (ent_idx, ent_level) = {
                 let pos = positions.get_mut(entity).unwrap();
                 (pos.idx, pos.level)
-
             };
             let level = dungeon.get_level_mut(ent_level).unwrap();
-            let delta = wants_to_move.idx as i32 - ent_idx as i32;
+            let delta = wants_to_move.idx - ent_idx;
             let ent_destination_idx = wants_to_move.idx;
             if let Some(grabbing) = grabbing {
                 let mut thing_pos = positions.get_mut(grabbing.thing).unwrap();
-                let thing_destination_idx = thing_pos.idx as i32 + delta;
-                let thing_destination_is_ent_position = thing_destination_idx == ent_idx as i32;
+                let thing_destination_idx = thing_pos.idx + delta;
+                let thing_destination_is_ent_position = thing_destination_idx == ent_idx;
                 let thing_destination_is_blocked =
-                    level_utils::tile_is_blocked(thing_destination_idx as i32, &level)
+                    level_utils::tile_is_blocked(thing_destination_idx, &level)
                         && !thing_destination_is_ent_position;
                 let thing_current_idx = thing_pos.idx;
                 let ent_destination_is_thing_position = ent_destination_idx == thing_current_idx;
                 let ent_destination_is_blocked =
-                    level_utils::tile_is_blocked(ent_destination_idx as i32, &level)
+                    level_utils::tile_is_blocked(ent_destination_idx, &level)
                         && !ent_destination_is_thing_position;
                 if !thing_destination_is_blocked && !ent_destination_is_blocked {
-                    thing_pos.idx = thing_destination_idx as usize;
+                    thing_pos.idx = thing_destination_idx;
                     if let Some(_) = blocks_tiles.get(grabbing.thing) {
-                        level.blocked[thing_current_idx as usize] = false;
-                        level.blocked[thing_destination_idx as usize] = true;
+                        level.blocked[thing_current_idx] = false;
+                        level.blocked[thing_destination_idx] = true;
                     }
                 }
             }
-            if !level_utils::tile_is_blocked(ent_destination_idx as i32, &level) {
+            if !level_utils::tile_is_blocked(ent_destination_idx, &level) {
                 if let Some(mut viewshed) = viewshed {
                     viewshed.dirty = true;
                 }
                 let mut ent_pos = positions.get_mut(entity).unwrap();
                 ent_pos.idx = wants_to_move.idx;
                 if let Some(_) = blocks_tiles.get(entity) {
-                    level.blocked[ent_idx as usize] = false;
-                    level.blocked[ent_destination_idx as usize] = true;
+                    level.blocked[ent_idx] = false;
+                    level.blocked[ent_destination_idx] = true;
                 }
                 moved
                     .insert(entity, EntityMoved {})
